darkpool-client: Refactor Arbitrum client into `DarkpoolImpl` trait
