+++ darkpool-client/src/client/mod.rs
@@ -14,9 +14,9 @@ use constants::{DEVNET_DEPLOY_BLOCK, MAINNET_DEPLOY_BLOCK, TESTNET_DEPLOY_BLOCK}
 use util::err_str;
 
 use crate::{
-    abi::Darkpool::DarkpoolInstance,
     constants::Chain,
     errors::{DarkpoolClientConfigError, DarkpoolClientError},
+    traits::DarkpoolImpl,
 };
 
 mod contract_interaction;
@@ -26,8 +26,6 @@ mod event_indexing;
 /// includes a signer derived from a raw private key, and a provider that
 /// connects to the RPC endpoint over HTTP.
 pub type RenegadeProvider = DynProvider;
-/// A darkpool contract instance
-pub type Darkpool = DarkpoolInstance<RenegadeProvider>;
 /// A darkpool call builder type
 pub type DarkpoolCallBuilder<'a, C> = CallBuilder<&'a DynProvider, C>;
 
@@ -82,38 +80,38 @@ impl DarkpoolClientConfig {
 /// The darkpool client, which provides a higher-level interface to the darkpool
 /// contract for Renegade-specific access patterns.
 #[derive(Clone)]
-pub struct DarkpoolClient {
+pub struct DarkpoolClient<D: DarkpoolImpl> {
     /// The darkpool contract instance
-    darkpool: Darkpool,
+    darkpool: D,
     /// The block number at which the darkpool was deployed
     deploy_block: BlockNumber,
 }
 
-impl DarkpoolClient {
+impl<D: DarkpoolImpl> DarkpoolClient<D> {
     /// Constructs a new darkpool client from the given configuration
     #[allow(clippy::needless_pass_by_value)]
     pub fn new(config: DarkpoolClientConfig) -> Result<Self, DarkpoolClientError> {
         let darkpool_address = config.get_darkpool_address()?;
         let provider = config.get_provider()?;
-        let darkpool = Darkpool::new(darkpool_address, provider);
+        let darkpool = D::new(darkpool_address, provider);
         let deploy_block = config.get_deploy_block();
 
         Ok(Self { darkpool, deploy_block })
     }
 
     /// Get a darkpool contract client
-    pub fn darkpool_client(&self) -> &Darkpool {
+    pub fn darkpool_client(&self) -> &D {
         &self.darkpool
     }
 
     /// Get an alloy address for the darkpool contract
     pub fn darkpool_addr(&self) -> Address {
-        *self.darkpool.address()
+        self.darkpool.address()
     }
 
     /// Get a reference to some underlying RPC client
     pub fn provider(&self) -> &RenegadeProvider {
-        self.darkpool_client().provider()
+        self.darkpool.provider()
     }
 
     /// Get the chain ID
